+++ sprint2_Toothless world/BE/be_sprint2/src/main/java/com/example/be_sprint2/model/auth/Account.java
@@ -8,7 +8,7 @@ import java.util.Set;
 public class Account {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
     private String username;
     private String password;
     @JsonBackReference
@@ -21,7 +21,7 @@ public class Account {
     public Account() {
     }
 
-    public Account(Long id, String username, String password, Set<AccountRole> accountRoles, User user) {
+    public Account(Integer id, String username, String password, Set<AccountRole> accountRoles, User user) {
         this.id = id;
         this.username = username;
         this.password = password;
@@ -29,11 +29,11 @@ public class Account {
         this.user = user;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
