+++ sprint2_Toothless world/BE/be_sprint2/src/main/java/com/example/be_sprint2/repository/IAccountRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.Optional;
 
-public interface IAccountRepository extends JpaRepository<Account,Long> {
+public interface IAccountRepository extends JpaRepository<Account,Integer> {
     @Query(value = "SELECT * FROM accounts as a WHERE a.username = :username", nativeQuery = true)
     Optional<Account> findByUsername(@Param("username") String username);
 }
