+++ sprint2_Toothless world/BE/be_sprint2/src/main/java/com/example/be_sprint2/repository/ICategoryRepository.java
@@ -4,5 +4,5 @@ import com.example.be_sprint2.model.product.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface ICategoryRepository extends JpaRepository<Category, Long> {
+public interface ICategoryRepository extends JpaRepository<Category, Integer> {
 }
