+++ sprint2_Toothless world/BE/be_sprint2/src/main/java/com/example/be_sprint2/repository/ISizeRepository.java
@@ -3,5 +3,5 @@ import com.example.be_sprint2.model.product.Size;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface ISizeRepository extends JpaRepository<Size, Long> {
+public interface ISizeRepository extends JpaRepository<Size, Integer> {
 }
